Move everything in .travis.yml from install to before_install
The Travis CI documentation doesn't give any explanation of what the
difference is between install and before_install. I would like for the
IDE, package, and library installation steps to be done before the
snapshot is taken but it's not documented when this happens so by doing
everything at the earliest stage I can make it more likely to happen.