Add carriage return to end of each report line
For some reason GitHub is refusing to display random job reports:

We can make this file beautiful and searchable if this error is
corrected: It looks like row 54 should actually have 19 columns, instead
of 8.

Even though I can see no difference in the format of that line or the
preceding line. For some reason if I save it with Windows style line
endings (crlf) and re-upload the file to GitHub it displays correctly so
I'm  formatting the report this way in hopes it will solve the issue.