Pull from remote report repository before pushing
If multiple jobs complete around the same time
publish_report_to_repository can fail because of history conflict. This
is caused by one job pushing to the repository after the other job has
cloned the repository but before it has pushed, causing a history
conflict. To avoid this I have added a ;pull before the commit (because
this is the last chance to do a pull without the chance of a conflict)
and just in case, another pull before the push. It should be extremely
rare that a push from one job happens between the pull and push of
another job.

Issue reported at
https://github.com/MCUdude/MegaCore/pull/42#issuecomment-301298593