Eliminate the use of sudo and use symbolic links for installation
Thanks to some advice from @majekw, I discovered that sudo was only necessary because of the use of /usr/local/share as the folder the Arduino IDE was installed to. By changing to the use of a subfolder of $HOME I was able to do without sudo. Hopefully this will decrease the build time.

I also am able to use symbolic links for installation, which makes it unnecessary to uninstall the IDE by moving it back to the version folder. This should be faster than using mv for installation and uninstallation.