Improve verbosity control
Previously, even with verbosity level set to 0 the log would be way too cluttered up with output that is not normally useful. At verbosity level 0 only the arduino compilation output should be shown. Overly verbose logs will cause the user to be forced to use the raw log and can even lead to the job being terminated due to the log exceeding the maximum length.

Closes https://github.com/per1234/arduino-ci-script/issues/10