Revert to previous verbosity settings
The user may want to change the verbosity of .travis.yml output so I should not assume those settings. Instead store the values and revert to them before exiting the function.

I removed the unnecessary calls to enable_verbosity and disable_verbosity in the private functions as verbosity control will already be handled in the public functions and extra calls will cause the loss of the saved verbosity values as well as changing the verbosity of the rest of the public function after the private function call.