Tell ShellCheck to ignore the issues in comment_report_link
The line that publishes the report has a lot of problems reported by ShellCheck but all my efforts to resolve them have resulted in it no longer working. I have never had any problems with this command as is. So I'm giving up for now and just telling ShellCheck to ignore it. I hope to properly resolve this eventually.