Disable the "all" Travis CI job
The job that tests install_ide "all" now usually exceeds the maximum job duration, causing the build to error. The only other option I could see is allowing failure for this job but it increases the build time a lot for something I usually won't bother to check. So I think it's better to just uncomment this line during testing of work that might affect the "all" functionality.