Automatically generate available IDE version list
Previously I had the list of available IDE versions hardcoded into the script and thus would need to update the script for every IDE release. This means that any project which contains a copy of the script will not automatically adjust to IDE releases without the script being updated. This also means that I either need to push a script release or the users of the script must use non-release versions of the script.

Closes https://github.com/per1234/arduino-ci-script/issues/16