check_library_properties: Add a parameter to set the maximum recursive search depth
The unbounded recursive search caused false positives so I removed it. But then I realized that for the authors of hardware packages, it would be useful to point the script at their libraries folder and have it check all the libraries, rather than needing to hard-code each library path. So now the script defaults to only checking the root of the specified folder, but can be configured to search recursively to the specified maximum depth.