Remove unit tests from repository
The invalid sketch unit tests cause problems when arduino-ci-script is bundled with any project installed to a folder the Arduino IDE scans for sketches. When a sketch is found that doesn't have a compliant name, the Arduino IDE attempts to display a confirmation dialog. In the headless environment of Travis CI builds, this causes the build to hang until timeout with absolutely no indication of what the problem is.

It's quite unfortunate to need to move the unit tests to a separate repository, since I prefer to commit unit tests along with the code that requires it. This will also make it difficult to run matching unit tests for any given point in the development history (would need to specify the arduino-ci-script-tests commit hash in the arduino-ci-script .travis.yml file, and update this each time the tests were modified).