check_library_structure: Note implications of stray sketches for Web Editor in error messages
Unlike the Arduino IDE, the Arduino Web Editor requires that all sketches bundled with a library be under the examples folder. If a library contains a sketch anywhere else, it can not be imported in the Arduino Web Editor. The sketches outside the examples folder will not be accessible as examples for the pre-installed libraries of the Library Manager index. I already have checks for stray sketches and misspelled examples folder name in check_library_structure, but their error messages didn't previously note the more serious implications of this for Arduino Web Editor.