Add all Microsoft Store releases to the IDE version blocklist
For the second time, the script has been broken by a tag used for a release specific to the Microsoft Store version of the Arduino IDE. Both times this happened, the tag has followed a consistent format. So I have generalized the regular expression to exclude all such tags in hopes this will prevent any future occurrences of this issue.